Fix imports, move test classes to appropriate packages, fix wording.
