+++ deployment/compute-cloud-sql/src/main/java/io/spine/examples/todolist/server/computecloudsql/ComputeCloudSqlServer.java
@@ -22,7 +22,7 @@ package io.spine.examples.todolist.server.computecloudsql;
 
 import com.google.common.annotations.VisibleForTesting;
 import io.spine.base.Production;
-import io.spine.examples.todolist.DbCredentials;
+import io.spine.examples.todolist.rdbms.DbCredentials;
 import io.spine.examples.todolist.rdbms.DbConnectionProperties;
 import io.spine.examples.todolist.rdbms.RdbmsStorageFactorySupplier;
 import io.spine.examples.todolist.server.Server;
