+++ deployment/local-my-sql/src/main/java/io/spine/examples/todolist/server/localmysql/LocalMySqlServer.java
@@ -22,7 +22,7 @@ package io.spine.examples.todolist.server.localmysql;
 
 import com.google.common.annotations.VisibleForTesting;
 import io.spine.base.Production;
-import io.spine.examples.todolist.DbCredentials;
+import io.spine.examples.todolist.rdbms.DbCredentials;
 import io.spine.examples.todolist.rdbms.DbConnectionProperties;
 import io.spine.examples.todolist.rdbms.DbUrlPrefix;
 import io.spine.examples.todolist.rdbms.RdbmsStorageFactorySupplier;
